Revisión de los vacíos: el radio salía con decimal, y faltaba un caso
Repaso de todo lo migrado en la rama. Dos cosas:

- `noNearbyBusinessesInCategory` declaraba su placeholder como `double`, así que
  imprimía «dentro de 5.0 km». Venía de antes, pero ahora queda al lado de un
  botón que dice «Ampliar a 10 km» y el desajuste canta. Placeholder a `int`.
- El banco de reservas comprobaba que el botón del comensal apaga su filtro,
  pero no el del dueño. Añadido; muere con la mutación correspondiente.

Verificado además: paridad de claves entre los tres .arb, ninguna referencia
viva a las ocho claves borradas, y que el vacío renderiza sin excepción en los
cuatro contenedores donde vive ahora (los dos SliverFillRemaining, ContentColumn
y el PageView de favoritos), también en apaisado estrecho.

+++ lib/ui/views/home/pages/foodly_main_page/foodly_categories/categories_page.dart
@@ -337,7 +337,7 @@ class VacioDeCategoria extends StatelessWidget {
     return FoodlyEmptyView(
       intent: FoodlyEmptyIntent.filtro,
       title: S.current.categoryEmptyTitle,
-      subtitle: S.current.noNearbyBusinessesInCategory(radioActual),
+      subtitle: S.current.noNearbyBusinessesInCategory(radioActual.toInt()),
       actionLabel: puedeAmpliar ? S.current.widenRadius(siguiente.toInt()) : null,
       onAction: puedeAmpliar ? () => onAmpliarRadio(siguiente) : null,
     );
